requirements.txt
pandas==2.2.2
google-cloud-storage==2.16.0
google-cloud-bigquery==3.23.0
